+++ R/imports.R
@@ -4,7 +4,7 @@ NULL
 #' @importFrom sna gplot as.sociomatrix.sna
 NULL
 
-#' @importFrom igraph graph_from_adjacency_matrix set_vertex_attr distances
+#' @importFrom igraph graph_from_adjacency_matrix set_vertex_attr
 #'  any_multiple graph_attr_names as_adj is.loop set_graph_attr
 #' @importFrom network as.edgelist is.multiplex is.directed has.loops
 NULL